Renames variants; replicate variant logic
