+++ src/fronts/opinion/variantB/Collections.tsx
@@ -1,11 +1,12 @@
 import React from "react";
 import { Collection as ICollection } from "../../../api";
 import { TableRowCell } from "../../../layout/Table";
-import { GenericCollection } from "../../../collections/GenericCollection";
-import { InstagramCollection } from "../../../collections/InstagramCollection";
-import { LinkCollection } from "./components/LinkCollection";
-import { getDesignType } from "../../../utils/getDesignType";
 import { CommentCollection } from "./components/CommentCollection";
+import { EditorialCollection } from "./components/EditorialCollection";
+import { MediaCollection } from "./components/MediaCollection";
+import { LinkCollection } from "./components/LinkCollection";
+import { DefaultCollection } from "./components/DefaultCollection";
+import { getDesignType } from "../../../utils/getDesignType";
 
 export const Collections: React.FC<{
     frontId: string;
@@ -17,6 +18,10 @@ export const Collections: React.FC<{
         const designType = getDesignType(content);
 
         switch (designType) {
+            case "editorial":
+                return (
+                    <EditorialCollection collection={collection} salt={salt} />
+                );
             case "comment":
                 return (
                     <CommentCollection
@@ -25,14 +30,8 @@ export const Collections: React.FC<{
                         salt={salt}
                     />
                 );
-            case "editorial":
-                return (
-                    <InstagramCollection collection={collection} salt={salt} />
-                );
             case "media":
-                return (
-                    <InstagramCollection collection={collection} salt={salt} />
-                );
+                return <MediaCollection collection={collection} salt={salt} />;
             case "link":
                 // Ignore 'Guardian Subscribe/Masterclasses' collection without using 'display name'
                 // Look at combination of content type (curated/backfill),
@@ -45,9 +44,11 @@ export const Collections: React.FC<{
                 }
 
                 return <LinkCollection collection={collection} salt={salt} />;
+            case "default":
+                return (
+                    <DefaultCollection collection={collection} salt={salt} />
+                );
         }
-
-        return <GenericCollection collection={collection} salt={salt} />;
     });
 
     return <TableRowCell>{renderedCollections}</TableRowCell>;
