+++ src/fronts/opinion/variantB/components/CommentCollection.tsx
@@ -2,10 +2,9 @@ import React from "react";
 import { Collection as ICollection } from "../../../../api";
 import { Heading } from "./../../../../components/Heading";
 import { Multiline } from "./../../../../components/Multiline";
-import { CommentCard } from "./../../../../components/cards/CommentCard";
+import { CommentCardB } from "../../../../components/cards/CommentCardB";
 import { Padding } from "../../../../layout/Padding";
-import { TableRowCell } from "../../../../layout/Table";
-import { palette } from "@guardian/src-foundations";
+import { Grid as CommentGrid } from "./Grid";
 
 const frontIdShouldShowCommentGridImages = (frontId: string): boolean => {
     if (frontId === "email/opinion") {
@@ -18,36 +17,68 @@ export const CommentCollection: React.FC<{
     frontId: string;
     collection: ICollection;
     salt: string;
-}> = ({ collection, salt }) => {
-    const lightGrey = palette.neutral[97];
-    const firstCollection = collection.backfill[0];
-    const rest = collection.backfill.slice(1);
+}> = ({ frontId, collection, salt }) => {
+    // TODO handle curated collections
 
+    const c0 = collection.backfill[0];
+    const c1 = collection.backfill[1];
+    const grid_2_5 = collection.backfill.slice(2, 6);
+    const c6 = collection.backfill[6];
+    const grid_7_8 = collection.backfill.slice(7, 9);
+    const c9 = collection.backfill[9];
+
+    // TODO
     return (
         <>
-            <TableRowCell tdStyle={{ backgroundColor: lightGrey }}>
-                <Padding px={12} />
-                <Multiline />
-                <Heading heading={collection.displayName} />
-                <CommentCard
-                    content={firstCollection}
-                    salt={salt}
-                    size="large"
-                    shouldShowProfileImage
-                />
-                <Padding px={12} />
-                {rest.map((story, index) => (
-                    <>
-                        <CommentCard
-                            content={story}
-                            salt={salt}
-                            size="small"
-                            shouldShowImage={false}
-                        />
-                        {index < rest.length - 1 && <Padding px={12} />}
-                    </>
-                ))}
-            </TableRowCell>
+            <Multiline topPadding />
+            <Heading heading={collection.displayName} />
+
+            <CommentCardB
+                content={c0}
+                salt={salt}
+                size={"large"}
+                shouldShowImage={false}
+            />
+            <Padding px={12} />
+
+            <CommentCardB
+                content={c1}
+                salt={salt}
+                size={"large"}
+                shouldShowImage={false}
+            />
+            <Padding px={12} />
+
+            <CommentGrid
+                content={grid_2_5}
+                salt={salt}
+                shouldShowGridImages={frontIdShouldShowCommentGridImages(
+                    frontId
+                )}
+            />
+
+            <CommentCardB
+                content={c6}
+                salt={salt}
+                size={"large"}
+                shouldShowImage={false}
+            />
+            <Padding px={12} />
+
+            <CommentGrid
+                content={grid_7_8}
+                salt={salt}
+                shouldShowGridImages={frontIdShouldShowCommentGridImages(
+                    frontId
+                )}
+            />
+
+            <CommentCardB
+                content={c9}
+                salt={salt}
+                size={"large"}
+                shouldShowImage={false}
+            />
         </>
     );
 };
