+++ src/fronts/opinion/variantB/components/DefaultCollection.tsx
