+++ src/fronts/opinion/variantB/components/EditorialCollection.tsx
