+++ src/fronts/opinion/variantB/components/LinkCollection.tsx
@@ -1,42 +1,24 @@
 import React from "react";
 import { Collection as ICollection } from "../../../../api";
-import { TableRowCell } from "../../../../layout/Table";
-import { Heading } from "../../../../components/Heading";
-import { Padding } from "../../../../layout/Padding";
-import { palette } from "@guardian/src-foundations";
-import { Multiline } from "../../../../components/Multiline";
-import { HeadlineCard } from "../../../../components/cards/HeadlineCard";
+import { Heading } from "./../../../../components/Heading";
+import { Multiline } from "./../../../../components/Multiline";
+import { LinkGrid as LinkGridB } from "./Grid";
 
 export const LinkCollection: React.FC<{
     collection: ICollection;
     salt: string;
-}> = ({ collection }) => {
-    const lightGrey = palette.neutral[97];
-    const white = palette.neutral[100];
+}> = ({ collection, salt }) => {
+    if (collection.curated.length < 1) {
+        return null;
+    }
+
     const content = collection.curated;
+
     return (
         <>
-            <Padding px={12} backgroundColor={lightGrey} />
-            <Multiline />
-            <Heading
-                heading={collection.displayName}
-                backgroundColor={lightGrey}
-            />
-            <TableRowCell tableStyle={{ backgroundColor: lightGrey }}>
-                {content.map((story, index) => (
-                    <>
-                        <HeadlineCard
-                            content={story}
-                            borderWidth="thin"
-                            borderColor={palette.opinion.main}
-                            backgroundColor={white}
-                            layout="expanded"
-                            showArrow
-                        />
-                        {index < content.length - 1 && <Padding px={12} />}
-                    </>
-                ))}
-            </TableRowCell>
+            <Multiline topPadding />
+            <Heading heading={collection.displayName} />
+            {content && <LinkGridB content={content} salt={salt} />}
         </>
     );
 };
