+++ src/fronts/opinion/variantB/components/MediaCollection.tsx
