+++ src/fronts/opinion/variantC/Collections.tsx
