+++ src/fronts/opinion/variantC/components/CommentCollection.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Collection as ICollection } from "../../../../api";
+import { Padding } from "../../../../layout/Padding";
 import { Heading } from "./../../../../components/Heading";
 import { Multiline } from "./../../../../components/Multiline";
-import { CommentCardB } from "../../../../components/cards/CommentCardB";
-import { Padding } from "../../../../layout/Padding";
+import { CommentCardC } from "../../../../components/cards/CommentCardC";
 import { Grid as CommentGrid } from "./Grid";
 
 const frontIdShouldShowCommentGridImages = (frontId: string): boolean => {
@@ -33,7 +33,7 @@ export const CommentCollection: React.FC<{
             <Multiline topPadding />
             <Heading heading={collection.displayName} />
 
-            <CommentCardB
+            <CommentCardC
                 content={c0}
                 salt={salt}
                 size={"large"}
@@ -41,7 +41,7 @@ export const CommentCollection: React.FC<{
             />
             <Padding px={12} />
 
-            <CommentCardB
+            <CommentCardC
                 content={c1}
                 salt={salt}
                 size={"large"}
@@ -57,7 +57,7 @@ export const CommentCollection: React.FC<{
                 )}
             />
 
-            <CommentCardB
+            <CommentCardC
                 content={c6}
                 salt={salt}
                 size={"large"}
@@ -73,7 +73,7 @@ export const CommentCollection: React.FC<{
                 )}
             />
 
-            <CommentCardB
+            <CommentCardC
                 content={c9}
                 salt={salt}
                 size={"large"}
