+++ src/fronts/opinion/variantC/components/DefaultCollection.tsx
