+++ src/fronts/opinion/variantC/components/EditorialCollection.tsx
