+++ src/fronts/opinion/variantC/components/MediaCollection.tsx
