+++ src/fronts/opinion/variantZ/Collections.tsx
@@ -1,12 +1,11 @@
 import React from "react";
 import { Collection as ICollection } from "../../../api";
 import { TableRowCell } from "../../../layout/Table";
-import { CommentCollection } from "./components/CommentCollection";
-import { EditorialCollection } from "./components/EditorialCollection";
-import { MediaCollection } from "./components/MediaCollection";
+import { GenericCollection } from "../../../collections/GenericCollection";
+import { InstagramCollection } from "../../../collections/InstagramCollection";
 import { LinkCollection } from "./components/LinkCollection";
-import { DefaultCollection } from "./components/DefaultCollection";
 import { getDesignType } from "../../../utils/getDesignType";
+import { CommentCollection } from "./components/CommentCollection";
 
 export const Collections: React.FC<{
     frontId: string;
@@ -18,10 +17,6 @@ export const Collections: React.FC<{
         const designType = getDesignType(content);
 
         switch (designType) {
-            case "editorial":
-                return (
-                    <EditorialCollection collection={collection} salt={salt} />
-                );
             case "comment":
                 return (
                     <CommentCollection
@@ -30,8 +25,14 @@ export const Collections: React.FC<{
                         salt={salt}
                     />
                 );
+            case "editorial":
+                return (
+                    <InstagramCollection collection={collection} salt={salt} />
+                );
             case "media":
-                return <MediaCollection collection={collection} salt={salt} />;
+                return (
+                    <InstagramCollection collection={collection} salt={salt} />
+                );
             case "link":
                 // Ignore 'Guardian Subscribe/Masterclasses' collection without using 'display name'
                 // Look at combination of content type (curated/backfill),
@@ -44,11 +45,9 @@ export const Collections: React.FC<{
                 }
 
                 return <LinkCollection collection={collection} salt={salt} />;
-            case "default":
-                return (
-                    <DefaultCollection collection={collection} salt={salt} />
-                );
         }
+
+        return <GenericCollection collection={collection} salt={salt} />;
     });
 
     return <TableRowCell>{renderedCollections}</TableRowCell>;
